Update OpenIDConnectClient.php5
Patch https://github.com/jumbojett/OpenID-Connect-PHP/commit/19a23a84110095188c0d45492e56bd2d4bdebb43 introduced a problem for our environment where $header->kid is undefined. The code dies on line 406 and is unable to even throw the exception on line 410 because of the missing field. I'm not sure if the proposed fix is the best solution, but it does may the code work again.