updated signature verification in authenticate method
Previous behaviour:
If RSA is available try to check signature. Fails, if jwks_uri is not provided.

However, providing JSON Web Keys is cumbersome; especially for test purposes it is nice to be able to bypass this.
New behaviour:
If jwks_uri is missing, an Exception is only thrown when FORCE_SIGNATURE_VERIFICATION is passed. Maybe we should make this default.