Fix OpenIDConnectClient.php for issue with Safari
Only with a proxy server, and only with the Safari web browser, the `$protocol` ends up being set to a list 'http, https' which is invalid and thus causes failure "Redirect URI not configured properly'.

With this fix, I'm not rewriting the complex ternary code (because I don't know if I can replicate the setup where these various headers are present. However, I can check if $protocol has a comma in it; and if so just default to 'https' for $protocol.