Fix for claims containing a null nonce.
Nonce is optional according to the open ID spec, so we shouldn't throw
an exception if a nonce is null (or not set) within the claims.

This was working previously but a change to how session keys are checked
here: https://github.com/jumbojett/OpenID-Connect-PHP/pull/251 meant
that the nonce check was now strict and `null !== false`.

This fixes by checking first that the nonce is set before checking it
matches the nonce in the session.